+++ CHANGELOG.md
@@ -20,6 +20,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Fix index API by @dennisvang in #637 (backward incompatible)
 - Fix listing metadata records after publish by @MarekSuchanek in #658
 - Fix saving settings by @MarekSuchanek in #657
+- Fix saving metadata schema (draft and publish) by @MarekSuchanek in #672
 
 ## [1.17.2]